requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
google-generativeai
openpyxl